+++ tests/test.js
@@ -23,8 +23,14 @@
 
 // support async/await tests in node6
 if(!require('semver').gte(process.version, '8.6.0')) {
-  require('babel-register')({
-    presets: ['node6-es6']
+  require('@babel/register')({
+    presets: [
+      [
+        '@babel/preset-env', {
+          targets: 'node 6'
+        }
+      ]
+    ]
   });
 }