+++ webpack.config.js
@@ -19,7 +19,6 @@ const outputs = [
   {
     entry: [
       // 'babel-polyfill' is very large, list features explicitly
-      'regenerator-runtime/runtime',
       'core-js/fn/array/from',
       'core-js/fn/array/includes',
       'core-js/fn/map',
@@ -72,9 +71,13 @@ outputs.forEach((info) => {
           use: {
             loader: 'babel-loader',
             options: {
-              presets: ['env'],
+              presets: ['@babel/preset-env'],
               plugins: [
-                ['transform-object-rest-spread', {useBuiltIns: true}]
+                [
+                  '@babel/plugin-proposal-object-rest-spread',
+                  {useBuiltIns: true}
+                ],
+                '@babel/plugin-transform-runtime'
               ]
             }
           }
@@ -96,6 +99,7 @@ outputs.forEach((info) => {
 
   // plain unoptimized unminified bundle
   const bundle = webpackMerge(common, {
+    mode: 'development',
     output: {
       path: path.join(__dirname, 'dist'),
       filename: info.filenameBase + '.js',
@@ -112,6 +116,7 @@ outputs.forEach((info) => {
 
   // optimized and minified bundle
   const minify = webpackMerge(common, {
+    mode: 'production',
     output: {
       path: path.join(__dirname, 'dist'),
       filename: info.filenameBase + '.min.js',
@@ -120,6 +125,7 @@ outputs.forEach((info) => {
     },
     devtool: 'cheap-module-source-map',
     plugins: [
+      /*
       new webpack.optimize.UglifyJsPlugin({
         //beautify: true,
         compress: {
@@ -130,6 +136,7 @@ outputs.forEach((info) => {
         },
         sourceMap: true
       })
+      */
     ]
   });
   if(info.library === null) {