+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   - *Note*: Using headless Chrome vs PhantomJS may cause newer JS features to
     slip into releases without proper support for older runtimes and browsers.
     Please report such issues and they will be addressed.
+- Update webpack and babel.
 
 ## 1.3.0 - 2019-01-04